Changed stripe fnts and country-currency config
